fix(warnings): 修复 55 个 bool-condition/non-bool-operand/for-in warnings
- non_bool_condition: 21 个（dynamic 在 if 条件加 == true 或 as）
- non_bool_operand: 18 个（dynamic 在 && 操作数加类型转换）
- for_in_of_invalid_type: 16 个（dynamic 在 for-in 加 as List）
31 文件修改，仅添加类型转换/比较

+++ lib/page/group/group_detail/remove_member_provider.dart
@@ -143,7 +143,7 @@ class RemoveMemberService {
       final g = await gRepo.findById(groupId);
       final gmRepo = GroupMemberRepo();
 
-      final sum = payload['user_id_sum'] ?? 0;
+      final sum = (payload['user_id_sum'] ?? 0) as int;
       final memberList = payload['member_list'] ?? <dynamic>[];
 
       final gData = {
